Fixbug : style tamagui home-oder-notification-favourite , delete : MyHeader.tsx
