+++ app/index.tsx
@@ -1,15 +1,17 @@
 import { StyleSheet } from 'react-native'
 import { SafeAreaView } from 'react-native-safe-area-context'
 import { Text } from 'tamagui'
+import BottomTabBar from '~/components/template/BottomTab'
 
 const index = (): JSX.Element => {
   return (
-    <SafeAreaView style={styles.container}>
-      <Text
-        textAlign="center"
-        color={'red'} fontSize={20}
-        fontWeight={'bold'}>FE Shoe Ecommerce</Text>
-    </SafeAreaView>
+    // <SafeAreaView style={styles.container}>
+    //   <Text
+    //     textAlign="center"
+    //     color={'red'} fontSize={20}
+    //     fontWeight={'bold'}>FE Shoe Ecommerce</Text>
+    // </SafeAreaView>
+    <BottomTabBar />
   )
 }
