+++ components/template/AccountScreen.tsx
@@ -1,11 +1,15 @@
 import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
-
+import { StyleSheet } from 'react-native'
+import { SafeAreaView } from 'react-native'
+import { Text } from 'tamagui'
 const AccountScreen = (): any => {
   return (
-    <View style={styles.container}>
-      <Text>Settings Screen</Text>
-    </View>
+    <SafeAreaView style={styles.container}>
+      <Text
+        textAlign="center"
+        fontSize={20}
+      >AccountScreen</Text>
+    </SafeAreaView>
   )
 }
