+++ components/template/OderScreen.tsx
@@ -1,13 +1,23 @@
 import React from 'react'
-import { Text, View } from 'react-native'
-
+import { StyleSheet } from 'react-native'
+import { SafeAreaView } from 'react-native'
+import { Text } from 'tamagui'
 const OderScreen = (): any => {
   return (
-    <View>
-      <Text>OderScreen</Text>
-    </View>
+    <SafeAreaView style={styles.container}>
+      <Text
+        textAlign="center"
+        fontSize={20}
+      >OderScreen</Text>
+    </SafeAreaView>
   )
 }
-
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    flex: 1,
+    justifyContent: 'center'
+  }
+})
 
 export default OderScreen
